협업시작, vo클래스의 일부 속성들의 타입을 변경함
1. Long > Integer로 변환
2. item_code > int를 String로 변환함

+++ orderplanning/src/main/java/com/overc1ock/domain/ProcurementPlanVO.java
@@ -10,8 +10,8 @@ import lombok.Data;
 @Data
 public class ProcurementPlanVO {
 	
-	private Long pp_code;
-	private Long item_code;
+	private Integer pp_code;
+	private Integer item_code;
 	private String item_name;
 	private String process;
 	private Date production_date;
